README.md created
The program allows to register the entry of a person and to consult how many people have tried to enter. 
A person can enter if he/she is of legal age, that is, if his/her ID type is different from Identity Card and if his/her ID number is in accordance with local government regulations (people whose ID has the penultimate even number can exit if the day of the month is an odd number, and vice versa).
The program has exception handling (custom exceptions) and automatic testing implemented with JUnit Jupiter.